Rename geometry mod to data
